REFACTOR: 검색바와 추천 검색어 목록 컴포넌트 분리
