+++ src/App.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import Search from './pages/Search';
+import Main from './pages/Main';
 import GlobalStyle from './styles/GlobalStyle';
 
 function App() {
   return (
     <>
       <GlobalStyle />
-      <Search />
+      <Main />
     </>
   );
 }
