+++ src/pages/Main.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import Search from '../components/Search';
+import { styled } from 'styled-components';
+
+export default function Main() {
+  return (
+    <Container>
+      <Title>
+        국내 모든 임상시험 검색하고 <br /> 온라인으로 참여하기
+      </Title>
+      <Search />
+    </Container>
+  );
+}
+
+const Container = styled.div`
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  padding: 80px 0;
+`;
+
+const Title = styled.h2`
+  font-size: 2.125rem;
+  font-weight: 700;
+  letter-spacing: -0.018em;
+  line-height: 1.6;
+  text-align: center;
+  margin-bottom: 42px;
+`;
